+++ packages/office-ui-fabric-react/src/components/Persona/PersonaCoin/PersonaCoin.tsx
@@ -1,9 +1,9 @@
-import { styled } from '../../Utilities';
+import { styled } from '../../../Utilities';
 import {
   IPersonaCoinProps,
   IPersonaCoinStyleProps,
   IPersonaCoinStyles
-} from './Persona.types';
+} from '../Persona.types';
 import { PersonaCoinBase } from './PersonaCoin.base';
 import { getStyles } from './PersonaCoin.styles';
