+++ packages/office-ui-fabric-react/src/components/Persona/PersonaCoin/index.ts
@@ -1,0 +1,2 @@
+export * from './PersonaCoin';
+export * from './PersonaCoin.base';
