+++ packages/office-ui-fabric-react/src/components/Persona/PersonaPresence/PersonaPresence.tsx
@@ -1,9 +1,9 @@
-import { styled } from '../../Utilities';
+import { styled } from '../../../Utilities';
 import {
   IPersonaPresenceProps,
   IPersonaPresenceStyleProps,
   IPersonaPresenceStyles
-} from './Persona.types';
+} from '../Persona.types';
 import { PersonaPresenceBase } from './PersonaPresence.base';
 import { getStyles } from './PersonaPresence.styles';
