+++ packages/office-ui-fabric-react/src/components/Persona/PersonaPresence/index.ts
@@ -1,0 +1,2 @@
+export * from './PersonaPresence';
+export * from './PersonaPresence.base';
